# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Gestión de Pedidos de Aberturas",
        "description": "API REST para gestionar los pedidos de una empresa de aberturas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrarse como cliente",
                "description": "Crea un usuario nuevo con rol de usuario operativo y tipo cliente, y devuelve un token de acceso.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "locality_id",
                                    "phone_1",
                                    "phone_2"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Juan",
                                        "maxLength": 100
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Pérez",
                                        "maxLength": 100
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@ejemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Cliente123",
                                        "minLength": 8
                                    },
                                    "locality_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "phone_1": {
                                        "type": "string",
                                        "example": "3454111111",
                                        "maxLength": 30
                                    },
                                    "phone_2": {
                                        "type": "string",
                                        "example": "3454222222",
                                        "maxLength": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente."
                    },
                    "422": {
                        "description": "Los datos proporcionados no son válidos."
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Verifica email y contraseña y devuelve un token de acceso de Sanctum.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "superadmin@aberturas.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "SuperAdmin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso."
                    },
                    "401": {
                        "description": "Las credenciales proporcionadas son incorrectas."
                    },
                    "403": {
                        "description": "El usuario está desactivado."
                    },
                    "422": {
                        "description": "Los datos proporcionados no son válidos."
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Ver mis datos",
                "description": "Devuelve los datos del usuario dueño del token, con su rol, tipo y perfil.",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Datos del usuario obtenidos correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Elimina el token con el que se hizo el pedido. Después de esto, ese token deja de servir.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación generado por Laravel Sanctum.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}